友達検索機能エラー修正

+++ database/migrations/2020_02_20_111527_create_comments_table.php
@@ -15,7 +15,7 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->Integer('article_id')->unsigned();
+            $table->bigInteger('article_id')->unsigned();
             $table->text('body');
             $table->timestamps();
             //外部キーの設定
@@ -27,7 +27,7 @@ class CreateCommentsTable extends Migration
      * Reverse the migrations.
      *
      * @return void
-     */
+     */ 
     public function down()
     {
         Schema::dropIfExists('comments');
